v3 48 Move nav into its own module
